bug-fix: fixed rendering issues
Removed the loggedin route from the swipeable routes component
this was removed bc swipeable routes renders the component
onto the page before we need it. This was causing vertical scroll
along with data for the component to be loaded to the dom.

Moving the component outside of the swipeable routes fixed the pre-mature rendeing and
also fixed the vertical scoll that was on the login and signup pages